♻️ Move app icon fetching to system_utils
- Move steps from Preview to Detailed Activity
- Move getWorkoutIcon to getInstalledAppIcon in system_utils.dart
