+++ lib/data/model/activity.dart
@@ -13,10 +13,11 @@ interface class Data<T> {
 interface class Activity extends ActivityPreview {
   final List<Data<int>>? heartRates;
   final List<Data<double>>? speed;
+  final int steps;
 
   Activity({
     required super.distance,
-    required super.steps,
+    required this.steps,
     required super.activityType,
     required super.caloriesBurnt,
     required super.start,
@@ -31,7 +32,6 @@ interface class ActivityPreview {
   final HealthWorkoutActivityType activityType;
   final int caloriesBurnt;
   final double distance;
-  final int steps;
   final DateTime start;
   final DateTime end;
   final String sourceId;
@@ -41,7 +41,6 @@ interface class ActivityPreview {
     required this.activityType,
     required this.caloriesBurnt,
     required this.distance,
-    required this.steps,
     required this.start,
     required this.end,
     required this.sourceId,
@@ -58,7 +57,6 @@ interface class ActivityPreview {
       activityType: HealthWorkoutActivityType.WALKING,
       caloriesBurnt: json['caloriesBurnt'],
       distance: json['distance'],
-      steps: json['steps'],
       start: DateTime.parse(json['start']),
       end: DateTime.parse(json['end']),
       sourceId: json["sourceId"],
@@ -70,7 +68,6 @@ interface class ActivityPreview {
       'activityType': "",
       'caloriesBurnt': caloriesBurnt,
       'distance': distance,
-      'steps': steps,
       'start': start.toIso8601String(), // Convert DateTime to ISO8601 string
       'end': end.toIso8601String(), // Convert DateTime to ISO8601 string
     };
